tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/2d.canvas.readonly.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.readonly</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.canvas.readonly</h1>
<p class="desc">canvas is readonly</p>


<script>
var t = async_test("canvas is readonly");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var offscreenCanvas2 = new OffscreenCanvas(100, 50);
var d = ctx.canvas;
_assertDifferent(offscreenCanvas2, d, "offscreenCanvas2", "d");
ctx.canvas = offscreenCanvas2;
_assertSame(ctx.canvas, d, "ctx.canvas", "d");

t.done();

});
</script>
